requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.3
django-extensions==2.1.9
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
ipython==7.6.1
ipython-genutils==0.2.0
jedi==0.14.0
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
wcwidth==0.1.7
whitenoise==4.1.2
